refactor select for TeamMemberIntegrationAuth
