+++ packages/server/dataloader/gcalLoaders.ts
@@ -1,16 +1,12 @@
 import DataLoader from 'dataloader'
 import GcalOAuth2Manager from '../integrations/gcal/GcalOAuth2Manager'
-import {IGetTeamMemberIntegrationAuthQueryResult} from '../postgres/queries/generated/getTeamMemberIntegrationAuthQuery'
 import upsertTeamMemberIntegrationAuth from '../postgres/queries/upsertTeamMemberIntegrationAuth'
+import {TeamMemberIntegrationAuth} from '../postgres/types'
 import sendToSentry from '../utils/sendToSentry'
 import RootDataLoader from './RootDataLoader'
 
 export const freshGcalAuth = (parent: RootDataLoader) => {
-  return new DataLoader<
-    {teamId: string; userId: string},
-    IGetTeamMemberIntegrationAuthQueryResult | null,
-    string
-  >(
+  return new DataLoader<{teamId: string; userId: string}, TeamMemberIntegrationAuth | null, string>(
     async (keys) => {
       const results = await Promise.allSettled(
         keys.map(async ({teamId, userId}) => {
@@ -48,7 +44,7 @@ export const freshGcalAuth = (parent: RootDataLoader) => {
             await upsertTeamMemberIntegrationAuth(newGcalAuth)
             return newGcalAuth
           }
-          return gcalAuth as IGetTeamMemberIntegrationAuthQueryResult
+          return gcalAuth
         })
       )
       const vals = results.map((result) => (result.status === 'fulfilled' ? result.value : null))
