+++ packages/server/graphql/mutations/helpers/createAzureTask.ts
@@ -1,12 +1,12 @@
 import {DataLoaderWorker} from '../../../graphql/graphql'
-import {IGetTeamMemberIntegrationAuthQueryResult} from '../../../postgres/queries/generated/getTeamMemberIntegrationAuthQuery'
 import {IntegrationProviderAzureDevOps} from '../../../postgres/queries/getIntegrationProvidersByIds'
+import {TeamMemberIntegrationAuth} from '../../../postgres/types'
 import AzureDevOpsServerManager from '../../../utils/AzureDevOpsServerManager'
 
 const createAzureTask = async (
   rawContentStr: string,
   serviceProjectHash: string,
-  azureAuth: IGetTeamMemberIntegrationAuthQueryResult,
+  azureAuth: TeamMemberIntegrationAuth,
   dataLoader: DataLoaderWorker
 ) => {
   const provider = await dataLoader.get('integrationProviders').loadNonNull(azureAuth.providerId)
