--- packages/server/postgres/queries/src/getTeamMemberIntegrationAuthQuery.sql
@@ -1,8 +1,0 @@
-/*
- @name getTeamMemberIntegrationAuthQuery
- */
-SELECT * FROM "TeamMemberIntegrationAuth"
-WHERE "teamId" = :teamId
-AND "userId" = :userId
-AND "service" = :service
-AND "isActive" = TRUE;
